change yaml losses to fit smp longtype

+++ losses/boundary_loss.py
@@ -20,7 +20,8 @@ def one_hot(label, n_classes, requires_grad=True):
     """
     device = label.device
     one_hot_label = torch.eye(
-        n_classes, device=device, requires_grad=requires_grad)[label]
+        n_classes, device=device, requires_grad=requires_grad
+    )[label.type(torch.long)]
     one_hot_label = one_hot_label.transpose(1, 3).transpose(2, 3)
 
     return one_hot_label
